feat(rov): redesignn structure, finish rov.get()
add depth, gyro data update in rov.get()
