--- client/rovClient/__init__.py
@@ -1,1 +1,0 @@
-from .core import Rov
